feat: clean up code, remove duplication and add examples for streaming with async or not
