+++ posthog/ai/__init__.py
@@ -1,3 +1,4 @@
-from .openai import OpenAI, AsyncOpenAI
+from .providers.openai.openai import OpenAI
+from .providers.openai.openai_async import AsyncOpenAI
 
 __all__ = ["OpenAI", "AsyncOpenAI"]
